перенес обработку esc в modal
